Fix snapshot flag checks and coord prototype

+++ Quake/sv_main.c
@@ -775,7 +775,7 @@ static qboolean SV_SnapshotTossMoving (const edict_t *ent)
 {
 	float tinyvel = net_snap_tinyvel.value;
 
-	if (!(ent->v.flags & FL_ONGROUND))
+	if (!((int)ent->v.flags & FL_ONGROUND))
 		return true;
 	return VectorLengthSquared (ent->v.velocity) > (tinyvel * tinyvel);
 }
@@ -804,9 +804,9 @@ static byte SV_SnapshotFlagsForEnt (const edict_t *ent)
 
 static qboolean SV_SnapshotMandatoryEnt (const edict_t *ent)
 {
-	if (ent->v.flags & FL_CLIENT)
+	if ((int)ent->v.flags & FL_CLIENT)
 		return true;
-	if (ent->v.movetype == MOVETYPE_PUSH && SV_SnapshotMoverMoving (ent))
+	if ((int)ent->v.movetype == MOVETYPE_PUSH && SV_SnapshotMoverMoving (ent))
 		return true;
 	return false;
 }
